feat: introduce evaluation module and add structure generation and answer quality experiments
