+++ evaluate/experiments/single_turn/transfer_to_agent_exp.py
@@ -2,7 +2,7 @@ import time
 
 from opik import evaluate, Opik
 
-from evaluate.base import evaluation_task
+from evaluate.base.evaluation import evaluation_task
 from evaluate.constant import TRANSFER_TO_AGENT_QUALITY, MATMASTER_SUBAGENT
 from evaluate.metric.transfer_to_agent_quality import TransferToAgentQuality
